Moved SlackSearch into the slack services directory.
